Refactor done, still packaging
