+++ pyproject.toml
@@ -1,0 +1,22 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "risk_adjustment_model"
+version = "0.1.0"
+authors = [
+  { name="Phil Fehlinger", email="pjfehlinger@gmail.com" },
+]
+description = "Python implementation of healthcare risk adjustment models"
+readme = "README.md"
+requires-python = ">=3.8"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+]
+
+[project.urls]
+Homepage = "https://github.com/pfehlinger/risk_adjustment_model"
+Issues = "https://github.com/pfehlinger/risk_adjustment_model/issues"
