+++ risk_adjustment_model/medicare_model.py
@@ -3,9 +3,9 @@ import json
 import importlib.resources
 
 from pathlib import Path
-from src.v24 import age_sex_edits_v24, get_disease_interactions_v24, CMS_VARIABLES_V24
-from src.v28 import age_sex_edits_v28, get_disease_interactions_v28, CMS_VARIABLES_V28
-from src.utilities import determine_age
+from risk_adjustment_model.v24 import age_sex_edits_v24, get_disease_interactions_v24, CMS_VARIABLES_V24
+from risk_adjustment_model.v28 import age_sex_edits_v28, get_disease_interactions_v28, CMS_VARIABLES_V28
+from risk_adjustment_model.utilities import determine_age
 from typing import Union, Optional
 
 
@@ -587,7 +587,7 @@ class MedicareModel:
 
         """
         if not self.year:
-            with importlib.resources.path('src.reference_data', 'medicare') as data_dir:
+            with importlib.resources.path('reference_data', 'medicare') as data_dir:
                 dirs = os.listdir(data_dir / self.version)
                 years = [int(dir) for dir in dirs]
                 max_year = max(years)
@@ -612,7 +612,7 @@ class MedicareModel:
         Raises:
             FileNotFoundError: If the specified version directory or reference data directory does not exist.
         """
-        with importlib.resources.path('src.reference_data', 'medicare') as data_dir:
+        with importlib.resources.path('reference_data', 'medicare') as data_dir:
             data_directory = data_dir / self.version / str(self.model_year)
         
         return data_directory
