+++ risk_adjustment_model/v28.py
@@ -1,9 +1,3 @@
-import pandas as pd
-import logging
-import os
-import yaml
-
-from pathlib import Path
 
 CMS_VARIABLES_V28 = {
     'coding_intensity_adjuster': 0.059,
